[Feature 1][Style] Add style for deployment
