+++ src/ZodiacSign.jsx
@@ -20,15 +20,16 @@ const ZodiacSign = (props) => {
 
     // 띠 계산 (12로 나눈 나머지를 기준으로 띠를 판별)
     const zodiacIndex = birthYear % 12;
-    // console.log(zodiacIndex);
     const sign = zodiacSigns[zodiacIndex];
 
     return sign;
   };
 
   return (
-    <div>
-      <p>Your Zodiac Sign: {calculateZodiacSign()}</p>
+    <div className="Zodiac-Sign">
+      <p>
+        Your Zodiac Sign: <span>{calculateZodiacSign()}</span>
+      </p>
     </div>
   );
 };
